AIDis_model/requirements.txt
tensorflow>=2.12
pillow
matplotlib
numpy
opencv-python
scikit-learn